Clean hypervolume code
